Geo-spatially Modeling Dengue Epidemics in Urban Cities: a Case Study of Lahore, Pakistan
The study objective is to predict the epidemiological impact of dengue fever arbovirosis in urban tropical areas of Pakistan. To do so, we used the GPS-based data of the Aedes larvae collected during 2014-2015 in Lahore. We developed a Geographically Weighted Logistic Regression (GWLR) model for Geospatially predicting larvae presence or absence in Lahore. Data on rainfall, temperature are included along with time series of the Normalized Difference Vegetation Index (NDVI) derived from Landsat imagery. We observed a high spatial variability of the GWLR parameter estimates of these variables in the study area. The GWLR model significantly ($R_a^{2}$ = 0.78) explained the presence or absence of Aedes larvae with temperature, rainfall, and NDVI variables in South and Southeast of the study area. In the North and North-West, however, GWLR relationships were observed weak in highly populated areas. Interpolating GWLR coefficients generate more accurate maps of Aedes larvae presence or absence