Using Geographically Weighted Regression Kriging for crop yield mapping in West-Africa
Geographic information systems support application of statistical techniques to map spatially-referenced crop data. To do so optimally, errors and uncertainties need to be minimized that are often associated with operations on the data. This paper applies a spatial statistical approach to upscale crop yields from the field level towards the scale of Burkina Faso. Observed yields were related to the Normalized Difference Vegetation Index (NDVI) derived from SPOT-VEGETATION. The objective was to quantify the uncertainties at the subsequent steps. First, we applied a point pattern analysis to examine uncertainties due to the sampling network of field surveys in the country. Second, geographically weighted regression kriging (GWRK) was applied to upscale the yield observations and to quantify the corresponding uncertainty. The proposed method was demonstrated with the mapping of sorghum yields in Burkina Faso and results were compared with those from regression kriging (RK) and kriging with external drift using a local kriging neighborhood (KEDLN). The proposed method was validated with independent yield observations obtained from field surveys. We observed that the lower uncertainty range value increased with 39\%, and the upper uncertainty range value decreased with 51\%, when comparing GWRK with RK and KEDLN. Moreover, GWRK reduced the prediction error variance as compared to RK (20 versus 31) and to KEDLN (20 versus 39). We found that climate and topography had a major impact on the country’s sorghum yields. Further, the financial ability of farmers influenced the crop management, and, thus, the sorghum crop yields. We concluded that GWRK effectively utilized information present in the covariate datasets and improved accuracies of both the regional-scale mapping of sorghum yields and was able to quantify the associated uncertainty.